+++ src/txscan.C
@@ -432,7 +432,7 @@ auto Parser::scan_double_hat(codepoint c) -> bool {
         int hc4 = input_line[p + 8].hex_val();
         if (hc0 >= 0 && hc1 >= 0 && hc2 >= 0 && hc3 >= 0 && hc4 >= 0) {
             input_line_pos    = p + 8;
-            input_line[p + 8] = codepoint((hc0 << 16) + (hc1 << 12) + (hc2 << 8) + (hc3 << 4) + hc4);
+            input_line[p + 8] = codepoint(static_cast<unsigned>((hc0 << 16) + (hc1 << 12) + (hc2 << 8) + (hc3 << 4) + hc4));
             return true;
         }
     }
@@ -443,7 +443,7 @@ auto Parser::scan_double_hat(codepoint c) -> bool {
         int hc4 = input_line[p + 6].hex_val();
         if (hc1 >= 0 && hc2 >= 0 && hc3 >= 0 && hc4 >= 0) {
             input_line_pos    = p + 6;
-            input_line[p + 6] = codepoint((hc1 << 12) + (hc2 << 8) + (hc3 << 4) + hc4);
+            input_line[p + 6] = codepoint(static_cast<unsigned>((hc1 << 12) + (hc2 << 8) + (hc3 << 4) + hc4));
             return true;
         }
     }
@@ -452,7 +452,7 @@ auto Parser::scan_double_hat(codepoint c) -> bool {
         int hc2 = input_line[p + 2].hex_val();
         if (hc1 >= 0 && hc2 >= 0) {
             input_line_pos             = p + 2;
-            input_line[input_line_pos] = codepoint(16 * hc1 + hc2);
+            input_line[input_line_pos] = codepoint(static_cast<unsigned>(16 * hc1 + hc2));
             return true;
         }
     }
@@ -460,7 +460,7 @@ auto Parser::scan_double_hat(codepoint c) -> bool {
     if (!C.is_ascii()) return false;
     uchar c1 = C.value;
     input_line_pos++;
-    input_line[p + 1] = codepoint(c1 < 64 ? c1 + 64 : c1 - 64);
+    input_line[p + 1] = codepoint(static_cast<unsigned>(c1 < 64 ? c1 + 64 : c1 - 64));
     return true;
 }
 
@@ -836,7 +836,7 @@ void Parser::store_new_line(int n, bool vb) {
 void Parser::insert_endline_char() {
     input_line_pos = 0;
     int cc         = eqtb_int_table[endlinechar_code].get_val();
-    if (cc >= 0 && cc < int(nb_characters)) input_line.emplace_back(cc);
+    if (cc >= 0 && cc < int(nb_characters)) input_line.emplace_back(static_cast<unsigned>(cc));
 }
 
 // Like get_a_newline below. But we never print the line in the log file.
