+++ src/txtokenlist.h
@@ -131,7 +131,7 @@ class Mactab {
 private:
     Macro **table{nullptr};    // this contains the table
     int *   rc_table{nullptr}; // this contains the reference counts
-    int     cur_rc_mac_len{0}; // size of the table.
+    size_t  cur_rc_mac_len{0}; // size of the table.
     int     ptr{-1};           // pointer to the first free position
 private:
     void rc_mac_realloc();
